[FIX] #ISSUE-188 StockDataProvider plugin depends on issue
The hotfix for issue 188 makes extension attributes non functional for
stock. This plugin depends on sales channels being the only loaded
extension attribute for the stock class. As soon as that issue is
resolved and extension attributes begin to be used for stock, this
plugin would break the inventory admin ui. This commit fixes that
preemptively.