Remove erronious check
The isUseConfigManageStock variable only describes whether or not the stockItemConfiguration uses the default config value or the stockItemConfiguration specific config value for this the isManageStock value. It does not say anything about the value of the manage_stock setting. Since this logic is already correctly implemented in the getManageStock function this line can be completely removed.

As it was the line causes a bug where every PUT call to /products/{sku}/stockItems/{stock_item} endpoint with a use_config_manage_stock value of false will always cause that stockItem to register as in_stock until a reindex is run.

As it is right now this call will still set products in_stock when Manage stock is disabled. I don't know enough about that workflow to determine if that is the correct behavior.